# openapi.yaml
openapi: 3.0.3
info:
  title: Tourplanner
  description: |-
    This is the specification of the required API endpoints for the Tourplanner server.
  version: 1.0.0
servers:
  - url: http://localhost:10001
tags:
  - name: tour
    description: Tour creation and modification(CRUD)
  - name: tourlogs
    description: Tour Log creation and modification(CRUD)
  - name: map
    description: Maquest API
paths:
  /tour:
    post:
      tags:
        - tour
      summary: Create a new tour
      description: Create tour with all information
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#components/schemas/Tour'
        required: true
      responses:
        '201':
          description: Tour successfully created
        '409':
          description: Tour with same name already created
    get:
      tags:
        - tour
      summary: Get all existing tours
      description: Get all tours as a list
      responses:
        '200':
          description: Tour successfully retrieved
        '204':
          description: No tours found
  /tours:
    get:
      tags:
        - tour
      summary: Get all tours
      description: Get all tours from database
      responses:
        '200':
          description: Array with all tours
  /tour/{id}:
    put:
      tags:
        - tour
      summary: Updates the tour data for the given name.
      description: Retrieves the data for the tour provided in the route.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            example: 12
          description: The id from a tour.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#components/schemas/Tour'
        required: true
      responses:
        '200':
          description: Tour sucessfully updated.
        '404':
          description: Tour not found.
    delete:
      tags:
        - tour
      summary: Deletes an existing tour.
      description: Deletes an existing tour.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            example: 12
          description: The id of the tour to delete.
      responses:
        '200':
          description: Tour successfully deleted
        '404':
          description: The provided tour was not found.
  /tour/{id}/tourlog:
    post:
      tags:
        - tourlogs
      summary: Create a new tour log mapped to a tour
      description: Create tour log with all information
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            example: 12
          description: The id from a tour to which the tourlog should belong.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#components/schemas/Tourlog'
        required: true
      responses:
        '201':
          description: Tour log successfully created
        '404':
          description: The provided tourid was not found.
  /tourlogs/{id}:
    get:
      tags:
        - tourlogs
      summary: Get all tourlogs
      description: Get all tourlogs from a tour
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            example: 12
          description: The id from a tour to which the tourlogs belong
      responses:
        '200':
          description: Array with all tourlogs from tour
  /tourlog/{id}:
    put:
      tags:
        - tourlogs
      summary: Updates the tourlog data for the given id.
      description: Updates the tourlog data for the given id.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            example: 12
          description: The id from a tourlog.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#components/schemas/Tourlog'
        required: true
      responses:
        '200':
          description: Tourlog sucessfully updated.
        '404':
          description: Tourlog not found.
    delete:
      tags:
        - tourlogs
      summary: Deletes an existing tourlog.
      description: Deletes an existing tourlog.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            example: 12
          description: The id of the tourlog to delete.
      responses:
        '200':
          description: Tourlog successfully deleted
        '404':
          description: The provided tourlog was not found.
  /map:
    get:
      tags:
        - map
      summary: Fetch Map from Mapquest
      description: Image fetched from API
      parameters:
        - in: query
          name: start
          schema:
            type: string
          description: The starting location of a Route Request.
        - in: query
          name: end
          schema:
            type: string
          description: The ending location of a Route Request.
      responses:
        '200':
          description: Image(JPEG)
  /distance:
    get:
      tags:
        - map
      summary: Fetch Distance from Mapquest
      description: Distance and Time fetched from API
      parameters:
        - in: query
          name: from
          schema:
            type: string
          description: The starting location of a Route Request.
        - in: query
          name: to
          schema:
            type: string
          description: The ending location of a Route Request.
        - in: query
          name: routeType
          schema:
            type: string
          description: Specifies the type of route wanted. Acceptable values are fastest, shortest, pedestrian, bicycle
      responses:
        '200':
          description: Distance[miles] and Time[sec]


components:
  schemas:
    Tour:
      type: object
      properties:
        id:
          type: integer
          example: 12
        name:
          type: string
          example: Bergtour
        description:
          type: string
          example: Wanderung
        from:
          type: string
          example: Saalbach
        to:
          type: string
          example: Leogang
        type:
          type: string
          example: Wanderung
        distance:
          type: number
          example: 12,5
        time:
          type: integer
          example: 12321
        imgPath:
          type: string
          example: ./sfsdf/dsfd/dsf.img
    Tourlog:
      type: object
      properties:
        id:
          type: integer
          example: 12
        creationTime:
          type: object
          example: 2023-03-27T12:31:57+00:00
        comment:
          type: string
          example: Gute Tour
        difficulty:
          type: integer
          example: 2
        totalTime:
          type: integer
          example: 12321
        rating:
          type: integer
          example: 1



    